+++ pyproject.toml
@@ -25,6 +25,7 @@ include = ["CHANGELOG.md"]
 
 [tool.poetry.dependencies]
 python = "^3.6.1"
+ariadne = "^0.12.0"
 PyYAML = "^5.3"
 dotmap = "^1.3"
 fs = "^2.4"
@@ -47,7 +48,6 @@ ray = {extras = ["tune"], version = "^1.0.1", optional = true}
 deepdiff = {version = "^5.0.2", optional = true}
 pandas = {version = "^1.1.5", optional = true}
 tabulate = {version = "^0.8.7", optional = true}
-ariadne = {version = "^0.12.0", optional = true}
 uvicorn = {version = "^0.13.1", optional = true}
 dataset = {version = "^1.4.1", optional = true}
