+++ src/machinable/graphql/query/query_type.py
