+++ src/machinable/graphql/query/submission.py
@@ -1,0 +1,7 @@
+from machinable.graphql.query.query_type import query
+from machinable.submission.submission import Submission
+
+
+@query.field("submission")
+async def resolve_submission(obj, info, url):
+    return Submission(url)
