--- src/machinable/server/graphql/__init__.py
@@ -1,3 +1,0 @@
-from .directives import directives
-from .scalars import scalar_types
-from .types import mutation, query, subscription
