--- src/machinable/server/graphql/schema/query.graphql
@@ -1,5 +1,0 @@
-type Query {
-  server: String
-  submission(url: String!): Submission
-  submissionComponent(url: String!): SubmissionComponent
-}
