--- src/machinable/server/graphql/schema/scalars.graphql
@@ -1,2 +1,0 @@
-scalar JSON
-scalar DateTime
