--- src/machinable/server/graphql/schema/submissions.graphql
@@ -1,38 +1,0 @@
-type Submission {
-  unique_id: String
-  url: String
-  submission_id: String
-  seed: Float
-  timestamp: Float
-  code_backup: Boolean
-  code_diff: String
-  code_version: JSON @json
-  host: JSON @json
-  started_at: DateTime
-  finished_at: DateTime
-  is_finished: Boolean @method
-  is_started: Boolean @method
-  components: [SubmissionComponent]
-  submissions: [SubmissionComponent]
-  ancestor: Submission
-}
-
-type SubmissionComponent {
-  unique_id: String
-  url: String
-  component_id: String
-  config: JSON @json
-  flags: JSON @json
-  tuning: Boolean
-  components: [JSON] @json
-  host: JSON @json
-  state: JSON @json
-  started_at: DateTime
-  heartbeat_at: DateTime
-  finished_at: DateTime
-  is_finished: Boolean @method
-  is_started: Boolean @method
-  is_active: Boolean @method
-  is_incomplete: Boolean @method
-  submission: Submission
-}
