--- src/machinable/server/graphql/schema/subscription.graphql
@@ -1,8 +1,0 @@
-type IndexSubscriptionDelta {
-  experiments: [StorageExperiment]
-  updates: [StorageExperiment]
-}
-
-type Subscription {
-  index(index: JSON!, limit: Int): IndexSubscriptionDelta
-}
