--- src/machinable/server/graphql/types/__init__.py
@@ -1,3 +1,0 @@
-from .mutation import mutation
-from .query import query
-from .subscription import subscription
