--- src/machinable/server/graphql/types/query/__init__.py
@@ -1,8 +1,0 @@
-from machinable.server.graphql.query.query_type import query
-
-# resolvers
-from machinable.server.graphql.query.server import resolve_server
-from machinable.server.graphql.query.storage import (
-    resolve_storage_component,
-    resolve_storage_experiment,
-)
