--- src/machinable/server/graphql/types/subscription/__init__.py
@@ -1,8 +1,0 @@
-# resolver
-from machinable.server.graphql.subscription.index import (
-    index_generator,
-    index_resolver,
-)
-from machinable.server.graphql.subscription.subscription_type import (
-    subscription,
-)
