+++ src/machinable/server/server.py
@@ -1,27 +1,12 @@
 import os
 
-from ariadne import load_schema_from_path, make_executable_schema
 from ariadne.asgi import GraphQL
+from machinable.graphql import schema
 from machinable.server.filesystem import filesystem_resolver
-from machinable.server.graphql import (
-    directives,
-    mutation,
-    query,
-    scalar_types,
-    subscription,
-)
 from starlette.applications import Starlette
 from starlette.middleware import Middleware
 from starlette.middleware.cors import CORSMiddleware
 from starlette.routing import Route
-
-dir_path = os.path.dirname(os.path.realpath(__file__))
-schema = make_executable_schema(
-    load_schema_from_path(os.path.join(dir_path, "graphql/schema/")),
-    [*scalar_types, query, mutation, subscription],
-    directives=directives,
-)
-
 
 graphql = GraphQL(schema, keepalive=True)
